Major refactor into Utilities, Settings, and more
