+++ src/admin/class-assets.php
@@ -16,45 +16,31 @@ if ( ! class_exists( Assets::class ) ) {
 	class Assets {
 
 		/**
-		 * Get the Common instance.
+		 * The Common instance.
 		 *
 		 * @var Common
 		 */
 		private $common;
 
 		/**
-		 * Initialize the class and set its properties, with Common as a dependency.
-		 *
-		 * @param Common
+		 * Initialize the class and set its properties.
 		 */
-		public function __construct( Common $common ) {
-			$this->common = $common;
+		public function __construct() {
+			$this->common = new Common();
 		}
 
 		/**
 		 * Register the stylesheets for the admin area.
 		 */
 		public function enqueue_styles() {
-			wp_enqueue_style( $this->common->plugin_text_domain, plugin_dir_url( __FILE__ ) . 'css/style.css', [], $this->common->version, 'all' );
+			wp_enqueue_style( $this->common->plugin_text_domain(), plugin_dir_url( __FILE__ ) . 'css/style.css', [], $this->common->plugin_version(), 'all' );
 		}
 
 		/**
 		 * Register the JavaScript for the admin area.
 		 */
 		public function enqueue_scripts() {
-			/*
-			 * This function is provided for demonstration purposes only.
-			 *
-			 * An instance of this class should be passed to the run() function
-			 * defined in Loader as all of the hooks are defined
-			 * in that particular class.
-			 *
-			 * The Loader will then create the relationship
-			 * between the defined hooks and the functions defined in this
-			 * class.
-			 */
-
-			wp_enqueue_script( $this->common->plugin_text_domain, plugin_dir_url( __FILE__ ) . 'js/script.js', [ 'jquery' ], $this->common->version, false );
+			wp_enqueue_script( $this->common->plugin_text_domain(), plugin_dir_url( __FILE__ ) . 'js/script.js', [ 'jquery' ], $this->common->plugin_version(), false );
 		}
 	}
 }
