+++ src/admin/class-settings.php
@@ -3,33 +3,31 @@
 namespace WP_Plugin_Name\Admin;
 
 use WP_Plugin_Name as NS;
-use WP_Plugin_Name\Common\Common as Common;
+use WP_Plugin_Name\Common\Settings as Common_Settings;
 
 // Abort if this file is called directly.
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-if ( ! class_exists( Settings::class ) ) {
+if ( ! class_exists( Admin_Settings::class ) ) {
 	/**
 	 * The admin-specific settings.
 	 */
-	class Settings {
+	class Admin_Settings {
 
 		/**
-		 * Get the Common instance.
+		 * Get the Settings instance from Common.
 		 *
-		 * @var Common
+		 * @var Common_Settings
 		 */
-		private $common;
+		private $settings;
 
 		/**
-		 * Initialize the class and set its properties, with Common as a dependency.
-		 *
-		 * @param Common
+		 * Initialize the class and set its properties.
 		 */
-		public function __construct( Common $common ) {
-			$this->common = $common;
+		public function __construct() {
+			$this->settings = new Common_Settings();
 		}
 
 		/**
@@ -41,39 +39,10 @@ if ( ! class_exists( Settings::class ) ) {
 		 */
 		public function add_action_links( $links ) {
 			$mylinks = [
-				'<a href="' . esc_url( $this->get_main_settings_page_url() ) . '">' . $this->get_settings_word() . '</a>',
+				'<a href="' . esc_url( $this->settings->get_main_settings_page_url() ) . '">' . $this->settings->get_settings_word() . '</a>',
 			];
 
 			return array_merge( $mylinks, $links );
-		}
-
-		/**
-		 * The plugin's Settings page URL.
-		 *
-		 * @return string
-		 */
-		private function get_main_settings_page_url() {
-			$url = 'options-general.php?page=' . $this->get_settings_page_slug();
-
-			return admin_url( $url );
-		}
-
-		/**
-		 * The plugin's Settings page slug.
-		 *
-		 * @return string
-		 */
-		private function get_settings_page_slug() {
-			return $this->common->plugin_text_domain . '-settings';
-		}
-
-		/**
-		 * The translatable "Settings" text.
-		 *
-		 * @return string
-		 */
-		private function get_settings_word() {
-			return esc_html__( 'Settings', $this->common->plugin_text_domain );
 		}
 
 		/**
@@ -83,8 +52,8 @@ if ( ! class_exists( Settings::class ) ) {
 			add_options_page(
 				NS\get_plugin_display_name(),
 				NS\get_plugin_display_name(),
-				$this->common->required_capability(),
-				$this->get_settings_page_slug(),
+				$this->settings->common->required_capability(),
+				$this->settings->get_settings_page_slug(),
 				[ $this, 'settings_page' ]
 			);
 		}
@@ -93,26 +62,26 @@ if ( ! class_exists( Settings::class ) ) {
 		 * Outputs HTML for the plugin's Settings page.
 		 */
 		public function settings_page() {
-			if ( ! current_user_can( $this->common->required_capability() ) ) {
-				wp_die( esc_html__( 'You do not have sufficient permissions to access this page.', $this->common->plugin_text_domain ) );
+			if ( ! current_user_can( $this->settings->common->required_capability() ) ) {
+				wp_die( esc_html__( 'You do not have sufficient permissions to access this page.', $this->settings->common->plugin_text_domain() ) );
 			}
 
-			$link_to_customizer_panel = $this->common->get_link_to_customizer_panel();
+			$link_to_customizer_panel = $this->settings->get_link_to_customizer_panel();
 
 			?>
 			<div class="wrap">
-				<h1><?php echo NS\get_plugin_display_name() . ' ' . $this->get_settings_word();
+				<h1><?php echo NS\get_plugin_display_name() . ' ' . $this->settings->get_settings_word();
 					?></h1>
 
-				<p><?php esc_html_e( "This plugin uses the WordPress Customizer to set its options.", $this->common->plugin_text_domain ); ?></p>
-				<p><?php esc_html_e( "Click the button below to be taken directly to this plugin's section within the WordPress Customizer.", $this->common->plugin_text_domain ); ?></p>
+				<p><?php esc_html_e( "This plugin uses the WordPress Customizer to set its options.", $this->settings->common->plugin_text_domain() ); ?></p>
+				<p><?php esc_html_e( "Click the button below to be taken directly to this plugin's section within the WordPress Customizer.", $this->settings->common->plugin_text_domain() ); ?></p>
 				<p>
-					<?php esc_html_e( "TODO: Add more text here", $this->common->plugin_text_domain ); ?>
+					<?php esc_html_e( "TODO: Add more text here", $this->settings->common->plugin_text_domain() ); ?>
 				</p>
 				<p>
 					<a href="<?php echo esc_url( $link_to_customizer_panel ); ?>"
 					   class="button-primary">
-						<?php esc_html_e( 'Edit Plugin Settings in WP Customizer', $this->common->plugin_text_domain ) ?>
+						<?php esc_html_e( 'Edit Plugin Settings in WP Customizer', $this->settings->common->plugin_text_domain() ) ?>
 					</a>
 				</p>
 				<br><br>
